fixed some validation problems
previously due to this problems password was required when user tries to
update password of another user.
btw, need to rework validation by moving some of it to business logic
layer. Currently it's mostly at transport layer and furthermore - some
times validation of the same stuff repeats on different layers.