UserDTO.Public.DeletedAt type changed to *time.Time
This change will fix a problem when if deleted_at is NULL, it set to a
zero value of time.Time and this zero value is marshalled into a JSON,
so, for example: In response from /user/search endpoint there are always
will be a deletedAt field, but it will have a zero value of time.Time
(0001-01-01T00:00:00Z) if user isn't deleted. Now this field will be
omitted if user isn't deleted.